+++ skorch/net.py
@@ -664,6 +664,7 @@ class NeuralNet:
             * numpy arrays
             * torch tensors
             * pandas DataFrame or Series
+            * scipy sparse CSR matrices
             * a dictionary of the former three
             * a list/tuple of the former three
             * a Dataset
@@ -740,6 +741,7 @@ class NeuralNet:
             * numpy arrays
             * torch tensors
             * pandas DataFrame or Series
+            * scipy sparse CSR matrices
             * a dictionary of the former three
             * a list/tuple of the former three
             * a Dataset
@@ -785,6 +787,7 @@ class NeuralNet:
             * numpy arrays
             * torch tensors
             * pandas DataFrame or Series
+            * scipy sparse CSR matrices
             * a dictionary of the former three
             * a list/tuple of the former three
             * a Dataset
@@ -821,6 +824,7 @@ class NeuralNet:
             * numpy arrays
             * torch tensors
             * pandas DataFrame or Series
+            * scipy sparse CSR matrices
             * a dictionary of the former three
             * a list/tuple of the former three
             * a Dataset
@@ -871,6 +875,7 @@ class NeuralNet:
             * numpy arrays
             * torch tensors
             * pandas DataFrame or Series
+            * scipy sparse CSR matrices
             * a dictionary of the former three
             * a list/tuple of the former three
             * a Dataset
@@ -949,6 +954,7 @@ class NeuralNet:
             * numpy arrays
             * torch tensors
             * pandas DataFrame or Series
+            * scipy sparse CSR matrices
             * a dictionary of the former three
             * a list/tuple of the former three
             * a Dataset
@@ -985,6 +991,7 @@ class NeuralNet:
             * numpy arrays
             * torch tensors
             * pandas DataFrame or Series
+            * scipy sparse CSR matrices
             * a dictionary of the former three
             * a list/tuple of the former three
             * a Dataset
@@ -1017,6 +1024,7 @@ class NeuralNet:
             * numpy arrays
             * torch tensors
             * pandas DataFrame or Series
+            * scipy sparse CSR matrices
             * a dictionary of the former three
             * a list/tuple of the former three
             * a Dataset
@@ -1046,6 +1054,7 @@ class NeuralNet:
             * numpy arrays
             * torch tensors
             * pandas DataFrame or Series
+            * scipy sparse CSR matrices
             * a dictionary of the former three
             * a list/tuple of the former three
             * a Dataset
@@ -1098,6 +1107,7 @@ class NeuralNet:
             * numpy arrays
             * torch tensors
             * pandas DataFrame or Series
+            * scipy sparse CSR matrices
             * a dictionary of the former three
             * a list/tuple of the former three
             * a Dataset